slight rework. Still need to go over all the code and check the doc.
